test(prestashop): stabilize Jest suite vs OOM — worker caps + spec split (#976)
The self-hosted `test` CI job intermittently went red on
@openlinker/integrations-prestashop with zero failed assertions — a Jest
worker OOM-killed by the OS (signal=SIGKILL/exitCode=null) under the
full-suite `pnpm -r test` fan-out (every package's Jest runs concurrently,
each ~cores-1 workers). Allegro's allegro-http-client.spec.ts ballooned to
~950s in the same runs. A SIGKILL'd suite is visually identical to a real
failure, training reflexive re-runs (the #374 silent-failure trap).

Test-infra only — no libs/ or apps/ runtime code changes:

- Cap Jest workers + per-worker memory on the heavy prestashop + allegro
  packages: maxWorkers: 2 (absolute, not '50%' — deterministic regardless
  of runner core count) + workerIdleMemoryLimit: '512MB' (recycle a worker
  before the OS OOM-kills it).
- Bound cross-package fan-out: test:ci now runs `pnpm -r
  --workspace-concurrency=2 test`. pnpm's default is 4, so the bound must
  go below 4 to actually throttle concurrent heavy packages.
- Split the 2279-line prestashop-order-processor-manager.adapter.spec.ts
  (one Jest worker held all its state) into 4 per-area spec files
  (create-order, create-order-resolution, update-fulfillment,
  fulfillment-status) sharing setup via a new __tests__/mocks factory.
  Test bodies copied verbatim; total count unchanged (369 → 369).
- The factory is the sole owner of the cross-context
  CustomerProjectionRepositoryPort import; split specs read the mock off
  the typed harness (inferred). check-cross-context-imports.mjs allow-list
  updated in lockstep (spec path → factory path).
- docs/testing-guide.md Troubleshooting: document the OOM symptom, the
  caps, the --workspace-concurrency lever, and "don't reflexively re-run a
  red that's actually OOM".

lint, type-check, and the full unit suite pass; prestashop green across 10
local re-runs (the 20x-on-runner AC lands on this PR's CI). allegro suite
unchanged (510 tests, ~22s under the cap).
